각각의 컴포넌트를 별도 파일로 분리
